♻️ feat(hud): presence を app 所有の hud パネルへ移設（テレメトリ移設 第3弾）
__usketchPresence グローバルを排除。presenceStore と readPresenceMembers を
apps/web/src/lib/presence-store.ts へ切り出し、app-local プラグイン
createPresencePanelPlugin が ctx.hud.registerPanel で「Members」パネルを登録。

- HUD 側: members-panel.tsx・presence prop plumbing（plugin.tsx のグローバル読み取り／
  debug-hud の prop・MembersPanel 描画）・presence-types.ts を削除。
- app.tsx: モジュールスコープの presenceStore／global／readPresenceMembers／palette を
  削除し lib から import。awareness→presenceStore.set の effect は維持。
  createPresencePanelPlugin を base plugins に追加。
- メンバー一覧は Controls ドックの Members セクションに表示。
